% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gorpiper}
\alias{gorpiper}
\title{gorpipe-r}
\description{
R library for executing GOR queries using local GORpipe and loading data into R.
}
\section{gorpipe-r functions}{

The main function you'll be using is:
* \code{\link{get_gorpipe}} to set up a gorpipe object for executing GOR queries from R
}

\examples{
\dontrun{
library(gorpiper)

full path to gorpipe, e.g. 
path <- "~/User/gor-scripts/bin/gorpipe"

# Make gorpipe object
gorpipe <- get_gorpipe(path)

# Run a simple query

results <- gorpipe("gor #dbsnp# | top 10000")

results
}
}
